Do not enable duplicated bottom space when js is enabled
Current code enables an always-on padding-bottom: 9em for body
element to avoid body contents to be blocked by footer banner.

However when javascript is enabled, the body element receives
an extra margin-bottom that equals to the height of page__footer
element. This causes extra 9em of spaces at the bottom of each page.

This commit adds a line in the javascript resize monitor function,
and always set the padding-bottom for body element to be 0 on window
reload or window resizing. This eliminates the extra 9em space at
the bottom. In the meanwhile, the 9em space is kept when javascript
is disabled for the website, providing a bottom-line user experience
in the js-disabled case.